deps/v8/test/inspector/runtime/serialization-options-expected.txt

RemoteObject.deepSerializedValue

Running test: Object
testing expression: {key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}
Runtime.evaluate
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"deep"}
{
    id : <messageId>
    result : {
        result : {
            className : Object
            deepSerializedValue : {
                type : object
                value : [
                    [0] : [
                        [0] : key_level_1
                        [1] : {
                            type : object
                            value : [
                                [0] : [
                                    [0] : key_level_2
                                    [1] : {
                                        type : object
                                        value : [
                                            [0] : [
                                                [0] : key_level_3
                                                [1] : {
                                                    type : string
                                                    value : value_level_3
                                                }
                                            ]
                                        ]
                                    }
                                ]
                            ]
                        }
                    ]
                ]
            }
            description : Object
            objectId : <objectId>
            type : object
        }
    }
}
Runtime.callFunctionOn
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"deep"}
{
    id : <messageId>
    result : {
        result : {
            className : Object
            deepSerializedValue : {
                type : object
                value : [
                    [0] : [
                        [0] : key_level_1
                        [1] : {
                            type : object
                            value : [
                                [0] : [
                                    [0] : key_level_2
                                    [1] : {
                                        type : object
                                        value : [
                                            [0] : [
                                                [0] : key_level_3
                                                [1] : {
                                                    type : string
                                                    value : value_level_3
                                                }
                                            ]
                                        ]
                                    }
                                ]
                            ]
                        }
                    ]
                ]
            }
            description : Object
            objectId : <objectId>
            type : object
        }
    }
}
Runtime.evaluate
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"deep","maxDepth":0}
{
    id : <messageId>
    result : {
        result : {
            className : Object
            deepSerializedValue : {
                type : object
            }
            description : Object
            objectId : <objectId>
            type : object
        }
    }
}
Runtime.callFunctionOn
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"deep","maxDepth":0}
{
    id : <messageId>
    result : {
        result : {
            className : Object
            deepSerializedValue : {
                type : object
            }
            description : Object
            objectId : <objectId>
            type : object
        }
    }
}
Runtime.evaluate
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"deep","maxDepth":1}
{
    id : <messageId>
    result : {
        result : {
            className : Object
            deepSerializedValue : {
                type : object
                value : [
                    [0] : [
                        [0] : key_level_1
                        [1] : {
                            type : object
                        }
                    ]
                ]
            }
            description : Object
            objectId : <objectId>
            type : object
        }
    }
}
Runtime.callFunctionOn
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"deep","maxDepth":1}
{
    id : <messageId>
    result : {
        result : {
            className : Object
            deepSerializedValue : {
                type : object
                value : [
                    [0] : [
                        [0] : key_level_1
                        [1] : {
                            type : object
                        }
                    ]
                ]
            }
            description : Object
            objectId : <objectId>
            type : object
        }
    }
}
Runtime.evaluate
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"deep","maxDepth":2}
{
    id : <messageId>
    result : {
        result : {
            className : Object
            deepSerializedValue : {
                type : object
                value : [
                    [0] : [
                        [0] : key_level_1
                        [1] : {
                            type : object
                            value : [
                                [0] : [
                                    [0] : key_level_2
                                    [1] : {
                                        type : object
                                    }
                                ]
                            ]
                        }
                    ]
                ]
            }
            description : Object
            objectId : <objectId>
            type : object
        }
    }
}
Runtime.callFunctionOn
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"deep","maxDepth":2}
{
    id : <messageId>
    result : {
        result : {
            className : Object
            deepSerializedValue : {
                type : object
                value : [
                    [0] : [
                        [0] : key_level_1
                        [1] : {
                            type : object
                            value : [
                                [0] : [
                                    [0] : key_level_2
                                    [1] : {
                                        type : object
                                    }
                                ]
                            ]
                        }
                    ]
                ]
            }
            description : Object
            objectId : <objectId>
            type : object
        }
    }
}
Runtime.evaluate
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"deep","maxDepth":999}
{
    id : <messageId>
    result : {
        result : {
            className : Object
            deepSerializedValue : {
                type : object
                value : [
                    [0] : [
                        [0] : key_level_1
                        [1] : {
                            type : object
                            value : [
                                [0] : [
                                    [0] : key_level_2
                                    [1] : {
                                        type : object
                                        value : [
                                            [0] : [
                                                [0] : key_level_3
                                                [1] : {
                                                    type : string
                                                    value : value_level_3
                                                }
                                            ]
                                        ]
                                    }
                                ]
                            ]
                        }
                    ]
                ]
            }
            description : Object
            objectId : <objectId>
            type : object
        }
    }
}
Runtime.callFunctionOn
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"deep","maxDepth":999}
{
    id : <messageId>
    result : {
        result : {
            className : Object
            deepSerializedValue : {
                type : object
                value : [
                    [0] : [
                        [0] : key_level_1
                        [1] : {
                            type : object
                            value : [
                                [0] : [
                                    [0] : key_level_2
                                    [1] : {
                                        type : object
                                        value : [
                                            [0] : [
                                                [0] : key_level_3
                                                [1] : {
                                                    type : string
                                                    value : value_level_3
                                                }
                                            ]
                                        ]
                                    }
                                ]
                            ]
                        }
                    ]
                ]
            }
            description : Object
            objectId : <objectId>
            type : object
        }
    }
}
Runtime.evaluate
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"deep","maxDepth":999,"additionalParameters":{"foo":"bar","baz":"qux"}}
{
    id : <messageId>
    result : {
        result : {
            className : Object
            deepSerializedValue : {
                type : object
                value : [
                    [0] : [
                        [0] : key_level_1
                        [1] : {
                            type : object
                            value : [
                                [0] : [
                                    [0] : key_level_2
                                    [1] : {
                                        type : object
                                        value : [
                                            [0] : [
                                                [0] : key_level_3
                                                [1] : {
                                                    type : string
                                                    value : value_level_3
                                                }
                                            ]
                                        ]
                                    }
                                ]
                            ]
                        }
                    ]
                ]
            }
            description : Object
            objectId : <objectId>
            type : object
        }
    }
}
Runtime.callFunctionOn
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"deep","maxDepth":999,"additionalParameters":{"foo":"bar","baz":"qux"}}
{
    id : <messageId>
    result : {
        result : {
            className : Object
            deepSerializedValue : {
                type : object
                value : [
                    [0] : [
                        [0] : key_level_1
                        [1] : {
                            type : object
                            value : [
                                [0] : [
                                    [0] : key_level_2
                                    [1] : {
                                        type : object
                                        value : [
                                            [0] : [
                                                [0] : key_level_3
                                                [1] : {
                                                    type : string
                                                    value : value_level_3
                                                }
                                            ]
                                        ]
                                    }
                                ]
                            ]
                        }
                    ]
                ]
            }
            description : Object
            objectId : <objectId>
            type : object
        }
    }
}
Runtime.evaluate
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"deep","maxDepth":999,"additionalParameters":{"INCORRECT_ADDITIONAL_PARAMETER":{}}}
{
    error : {
        code : -32602
        message : Values of serializationOptions.additionalParameters can be only of type string or integer.
    }
    id : <messageId>
}
Runtime.callFunctionOn
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"deep","maxDepth":999,"additionalParameters":{"INCORRECT_ADDITIONAL_PARAMETER":{}}}
{
    error : {
        code : -32602
        message : Values of serializationOptions.additionalParameters can be only of type string or integer.
    }
    id : <messageId>
}
Runtime.evaluate
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"json"}
{
    id : <messageId>
    result : {
        result : {
            type : object
            value : {
                key_level_1 : {
                    key_level_2 : {
                        key_level_3 : value_level_3
                    }
                }
            }
        }
    }
}
Runtime.callFunctionOn
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"json"}
{
    id : <messageId>
    result : {
        result : {
            type : object
            value : {
                key_level_1 : {
                    key_level_2 : {
                        key_level_3 : value_level_3
                    }
                }
            }
        }
    }
}
Runtime.evaluate
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"json","maxDepth":1}
{
    id : <messageId>
    result : {
        result : {
            type : object
            value : {
                key_level_1 : {
                    key_level_2 : {
                        key_level_3 : value_level_3
                    }
                }
            }
        }
    }
}
Runtime.callFunctionOn
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"json","maxDepth":1}
{
    id : <messageId>
    result : {
        result : {
            type : object
            value : {
                key_level_1 : {
                    key_level_2 : {
                        key_level_3 : value_level_3
                    }
                }
            }
        }
    }
}
Runtime.evaluate
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"idOnly"}
{
    id : <messageId>
    result : {
        result : {
            className : Object
            description : Object
            objectId : <objectId>
            type : object
        }
    }
}
Runtime.callFunctionOn
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"idOnly"}
{
    id : <messageId>
    result : {
        result : {
            className : Object
            description : Object
            objectId : <objectId>
            type : object
        }
    }
}
Runtime.evaluate
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"INCORRECT_SERIALIZATION_TYPE"}
{
    error : {
        code : -32602
        message : Unknown serializationOptions.serialization value INCORRECT_SERIALIZATION_TYPE
    }
    id : <messageId>
}
Runtime.callFunctionOn
expression: "{key_level_1: {key_level_2: {key_level_3: 'value_level_3'}}}"
serializationOptions: {"serialization":"INCORRECT_SERIALIZATION_TYPE"}
{
    error : {
        code : -32602
        message : Unknown serializationOptions.serialization value INCORRECT_SERIALIZATION_TYPE
    }
    id : <messageId>
}
